+++ messages/en/motion.php
@@ -166,6 +166,8 @@ return [
     'support_finish_done'                => 'The motion is now officially submitted',
     'prefix_new_code'                    => 'new',
     'comment_for'                        => 'Comment for',
+    'share_facebook'                     => 'Share on Facebook',
+    'share_twitter'                      => 'Share on Twitter',
 
     'diff_bc'          => 'Changes',
     'diff_title'       => 'Changes from %FROM% to %TO%',