+++ models/sectionTypes/TabularData.php
@@ -136,7 +136,7 @@ class TabularData extends ISectionType
             return;
         }
 
-        if (!$pdfLayout->isSkippingSectionTitles($this->section)) {
+        if ($this->section->getSettings()->printTitle) {
             $pdfLayout->printSectionHeading($this->section->getSettings()->title);
         }