+++ models/sectionTypes/TextHTML.php
@@ -94,7 +94,7 @@ class TextHTML extends Text
             return;
         }
 
-        if (!$pdfLayout->isSkippingSectionTitles($this->section)) {
+        if ($this->section->getSettings()->printTitle) {
             $pdfLayout->printSectionHeading($this->section->getSettings()->title);
         }