+++ models/siteSpecificBehavior/DefaultBehavior.php
@@ -4,7 +4,6 @@ namespace app\models\siteSpecificBehavior;
 
 use app\components\MotionSorter;
 use app\models\db\Consultation;
-use app\models\db\IMotionSection;
 use app\models\policies\IPolicy;
 
 class DefaultBehavior
@@ -34,16 +33,6 @@ class DefaultBehavior
     public static function getSortedMotionsSort($prefix1, $prefix2)
     {
         return MotionSorter::getSortedMotionsSort($prefix1, $prefix2);
-    }
-
-    /**
-     * @param IMotionSection $section
-     * @return bool
-     * @SuppressWarnings(PHPMD.UnusedFormalParameter)
-     */
-    public static function showSectionIntroductionInPdf($section)
-    {
-        return true;
     }
 
     /**