+++ views/pdfLayouts/DBJR.php
@@ -3,7 +3,6 @@
 namespace app\views\pdfLayouts;
 
 use app\models\db\Amendment;
-use app\models\db\IMotionSection;
 use app\models\db\Motion;
 use setasign\Fpdi\TcpdfFpdi;
 use yii\helpers\Html;
@@ -170,18 +169,5 @@ class DBJR extends IPDFLayout
         $this->pdf = $pdf;
 
         return $pdf;
-    }
-
-    /**
-     * @param IMotionSection $section
-     * @return bool
-     */
-    public function isSkippingSectionTitles(IMotionSection $section)
-    {
-        if ($section->getSettings()->title == \Yii::t('motion', 'motion_text')) {
-            return true;
-        }
-        return false;
-
     }
 }