+++ History.md
@@ -4,6 +4,7 @@
 
 - When creating a motion, the confirmation page now shows a preview of the generated PDF.
 - Motion types can now force motion titles to have a certain beginning, like "Application: ".
+- For each section of a motion type, it is now possible to specify if the title will be explicitly printed in the PDF of not.
 
 ## Version 4.0.0 (2018-09-01)